Restructure code to ease changing the setup_py reader
This prepares for the next commit which will move to AST parsing the
setup.py.
