--- src/tomlize/loaders/__init__.py
@@ -1,5 +1,0 @@
-from . import setup_py
-
-__all__ = [
-    "setup_py",
-]
