+++ src/tomlize/setup_py/__init__.py
@@ -1,0 +1,3 @@
+from .transformer import extract
+
+__all__ = ["extract"]
